@@ -4,18 +4,18 @@ version: 0.2.0
 publish_to: none
 
 environment:
-  sdk: ^3.5.0
+  sdk: ^3.8.0
 
 dev_dependencies:
-  path:
-  pub_semver:
-  dev_build:
-  pedantic:
-  lints: '>=3.0.0'
+  path: ">=1.9.0"
+  pub_semver: ">=2.1.4"
+  dev_build: ">=1.1.2"
+  lints: ">=5.0.0"
   tekartik_lints:
     git:
       url: https://github.com/tekartik/common.dart
-      ref: dart2_3
+      ref: dart3a
       path: packages/lints
     version: '>=0.1.0'
+resolution: workspace
 
