@@ -19,17 +19,17 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData.dark(
-          // This is the theme of your application.
-          //
-          // Try running your application with "flutter run". You'll see the
-          // application has a blue toolbar. Then, without quitting the app, try
-          // changing the primarySwatch below to Colors.green and then invoke
-          // "hot reload" (press "r" in the console where you ran "flutter run",
-          // or simply save your changes to "hot reload" in a Flutter IDE).
-          // Notice that the counter didn't reset back to zero; the application
-          // is not restarted.
-          //prim: Colors.blue,
-          ),
+        // This is the theme of your application.
+        //
+        // Try running your application with "flutter run". You'll see the
+        // application has a blue toolbar. Then, without quitting the app, try
+        // changing the primarySwatch below to Colors.green and then invoke
+        // "hot reload" (press "r" in the console where you ran "flutter run",
+        // or simply save your changes to "hot reload" in a Flutter IDE).
+        // Notice that the counter didn't reset back to zero; the application
+        // is not restarted.
+        //prim: Colors.blue,
+      ),
       home: const TrackItemListPage(),
     );
   }
@@ -50,13 +50,7 @@ class TrackItemListPage extends StatefulWidget {
   State<TrackItemListPage> createState() => _TrackItemListPageState();
 }
 
-enum MenuAction {
-  clear,
-  runNow,
-  runIn15s,
-  runIn30s,
-  exit,
-}
+enum MenuAction { clear, runNow, runIn15s, runIn30s, exit }
 
 class _TrackItemListPageState extends State<TrackItemListPage> {
   final _itemList = BehaviorSubject<ItemList>();
@@ -78,8 +72,9 @@ class _TrackItemListPageState extends State<TrackItemListPage> {
           }
           _itemList.add(items);
           // Wait for modification
-          await service.onItemsUpdated
-              .firstWhere((element) => element != items.lastChangeId);
+          await service.onItemsUpdated.firstWhere(
+            (element) => element != items.lastChangeId,
+          );
         } catch (e) {
           print(e);
         }
@@ -133,11 +128,12 @@ class _TrackItemListPageState extends State<TrackItemListPage> {
           PopupMenuButton<MenuAction>(
             itemBuilder: (context) => [
               PopupMenuItem<MenuAction>(
-                  value: MenuAction.clear,
-                  child: const Text('Clear'),
-                  onTap: () {
-                    service.clearItems();
-                  }),
+                value: MenuAction.clear,
+                child: const Text('Clear'),
+                onTap: () {
+                  service.clearItems();
+                },
+              ),
               if (Platform.isAndroid) ...[
                 PopupMenuItem<MenuAction>(
                   value: MenuAction.runNow,
@@ -154,129 +150,126 @@ class _TrackItemListPageState extends State<TrackItemListPage> {
                   },
                 ),
                 PopupMenuItem<MenuAction>(
-                    value: MenuAction.runIn30s,
-                    onTap: () {
-                      _triggerIn(context, 30);
-                    },
-                    child: const Text('Trigger in 30s')),
+                  value: MenuAction.runIn30s,
+                  onTap: () {
+                    _triggerIn(context, 30);
+                  },
+                  child: const Text('Trigger in 30s'),
+                ),
               ],
               PopupMenuItem<MenuAction>(
-                  value: MenuAction.exit,
-                  onTap: () {
-                    exit(0);
-                  },
-                  child: const Text('Exit')),
+                value: MenuAction.exit,
+                onTap: () {
+                  exit(0);
+                },
+                child: const Text('Exit'),
+              ),
             ],
           ),
         ],
       ),
       body: StreamBuilder<ItemList>(
-          stream: _itemList,
-          // Center is a layout widget. It takes a single child and positions it
-          // in the middle of the parent.
-          builder: (context, snapshot) {
-            if (!snapshot.hasData) {
-              return const Center(
-                child: CircularProgressIndicator(),
-              );
-            }
-            var groupList = snapshot.data!.groups.reversed.toList();
-            if (groupList.isEmpty) {
-              return const ListTile(
-                title: Center(child: Text('no data yet')),
-              );
-            }
-            // devPrint('groups: ${groupList.length}');
-            var groupIds =
-                Set<int>.from(groupList.map((group) => group.groupId));
-            groupExpandedMap
-                .removeWhere((key, value) => !groupIds.contains(key));
-            return SingleChildScrollView(
-              child: ExpansionPanelList(
-                expansionCallback: (int index, bool isExpanded) {
-                  setState(() {
-                    groupExpandedMap[groupList[index].groupId] = !isExpanded;
-                  });
-                },
-                children: groupList
-                    .map((group) => ExpansionPanel(
-                          headerBuilder: (context, isExpanded) {
-                            var timestamp = group.items.first.anyTimestamp;
-                            var startTimestamp = DateTime.tryParse(timestamp)!;
-                            var lastTimestamp = DateTime.tryParse(
-                                group.items.last.anyTimestamp)!;
-                            var durationText = lastTimestamp
-                                .difference(startTimestamp)
-                                .toString();
-                            var dotIndex = durationText.lastIndexOf('.');
-                            if (dotIndex != -1) {
-                              durationText =
-                                  durationText.substring(0, dotIndex);
-                            }
+        stream: _itemList,
+        // Center is a layout widget. It takes a single child and positions it
+        // in the middle of the parent.
+        builder: (context, snapshot) {
+          if (!snapshot.hasData) {
+            return const Center(child: CircularProgressIndicator());
+          }
+          var groupList = snapshot.data!.groups.reversed.toList();
+          if (groupList.isEmpty) {
+            return const ListTile(title: Center(child: Text('no data yet')));
+          }
+          // devPrint('groups: ${groupList.length}');
+          var groupIds = Set<int>.from(groupList.map((group) => group.groupId));
+          groupExpandedMap.removeWhere((key, value) => !groupIds.contains(key));
+          return SingleChildScrollView(
+            child: ExpansionPanelList(
+              expansionCallback: (int index, bool isExpanded) {
+                setState(() {
+                  groupExpandedMap[groupList[index].groupId] = !isExpanded;
+                });
+              },
+              children: groupList
+                  .map(
+                    (group) => ExpansionPanel(
+                      headerBuilder: (context, isExpanded) {
+                        var timestamp = group.items.first.anyTimestamp;
+                        var startTimestamp = DateTime.tryParse(timestamp)!;
+                        var lastTimestamp = DateTime.tryParse(
+                          group.items.last.anyTimestamp,
+                        )!;
+                        var durationText = lastTimestamp
+                            .difference(startTimestamp)
+                            .toString();
+                        var dotIndex = durationText.lastIndexOf('.');
+                        if (dotIndex != -1) {
+                          durationText = durationText.substring(0, dotIndex);
+                        }
 
-                            var localDateTime =
-                                startTimestamp.toLocal().toIso8601String();
-                            var dateTimeText = localDateTime
-                                .substring(0, 19)
-                                .replaceAll('T', ' ');
+                        var localDateTime = startTimestamp
+                            .toLocal()
+                            .toIso8601String();
+                        var dateTimeText = localDateTime
+                            .substring(0, 19)
+                            .replaceAll('T', ' ');
 
-                            var subtitle =
-                                'Duration $durationText (${group.items.length} actions)';
-                            return ListTile(
-                              leading: Column(
-                                crossAxisAlignment: CrossAxisAlignment.center,
-                                mainAxisAlignment: MainAxisAlignment.center,
-                                children: [
-                                  Text(
-                                    group.items.first.tag.v ?? 'back',
-                                    textAlign: TextAlign.center,
-                                    style: const TextStyle(
-                                        fontWeight: FontWeight.bold),
-                                  ),
-                                  // Text(item.first.groupId.v?.toString() ?? ''),
-                                ],
-                              ),
-                              title: Text(dateTimeText),
-                              subtitle: Text(subtitle),
-                            );
-                          },
-                          body: Column(
-                              children: group.items.map(
-                            (item) {
-                              String dateTimeText;
-                              try {
-                                dateTimeText = DateTime.parse(item.anyTimestamp)
-                                    .toLocal()
-                                    .toIso8601String()
-                                    .substring(11, 19);
-                              } catch (_) {
-                                dateTimeText = '<none>';
-                              }
-                              var sb = StringBuffer();
-                              sb.add(dateTimeText);
-
-                              var title = sb.toString();
-                              sb = StringBuffer();
-                              sb
-                                ..add('pid: ${item.processId.v}')
-                                ..add('isolate: ${item.isolateName.v}');
-                              var subtitle = sb.toString();
-                              return ListTile(
-                                dense: true,
-                                leading: const SizedBox(
-                                  width: 32,
+                        var subtitle =
+                            'Duration $durationText (${group.items.length} actions)';
+                        return ListTile(
+                          leading: Column(
+                            crossAxisAlignment: CrossAxisAlignment.center,
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            children: [
+                              Text(
+                                group.items.first.tag.v ?? 'back',
+                                textAlign: TextAlign.center,
+                                style: const TextStyle(
+                                  fontWeight: FontWeight.bold,
                                 ),
-                                title: Text(title),
-                                subtitle: Text(subtitle),
-                              );
-                            },
-                          ).toList()),
-                          isExpanded: groupExpandedMap[group.groupId] ??= false,
-                        ))
-                    .toList(),
-              ),
-            );
-          }),
+                              ),
+                              // Text(item.first.groupId.v?.toString() ?? ''),
+                            ],
+                          ),
+                          title: Text(dateTimeText),
+                          subtitle: Text(subtitle),
+                        );
+                      },
+                      body: Column(
+                        children: group.items.map((item) {
+                          String dateTimeText;
+                          try {
+                            dateTimeText = DateTime.parse(
+                              item.anyTimestamp,
+                            ).toLocal().toIso8601String().substring(11, 19);
+                          } catch (_) {
+                            dateTimeText = '<none>';
+                          }
+                          var sb = StringBuffer();
+                          sb.add(dateTimeText);
+
+                          var title = sb.toString();
+                          sb = StringBuffer();
+                          sb
+                            ..add('pid: ${item.processId.v}')
+                            ..add('isolate: ${item.isolateName.v}');
+                          var subtitle = sb.toString();
+                          return ListTile(
+                            dense: true,
+                            leading: const SizedBox(width: 32),
+                            title: Text(title),
+                            subtitle: Text(subtitle),
+                          );
+                        }).toList(),
+                      ),
+                      isExpanded: groupExpandedMap[group.groupId] ??= false,
+                    ),
+                  )
+                  .toList(),
+            ),
+          );
+        },
+      ),
       floatingActionButton: FloatingActionButton(
         onPressed: _workOnce,
         tooltip: 'Worn once',
